Removed ThrowHelper; Moved his methods to corresponding tokens; add check exists value in dictionary
