+++ cs/Markdown/TokenModels/ItalicToken.cs
@@ -1,4 +1,5 @@
-﻿using Markdown.Core;
+﻿using System;
+using Markdown.Core;
 
 namespace Markdown.TokenModels
 {
@@ -19,7 +20,10 @@ namespace Markdown.TokenModels
 
         public static ItalicToken Create(string mdString, int startIndex)
         {
-            var endIndex = GetEndOfToken(mdString, startIndex);
+            var analyzer = new StringAnalyzer(mdString);
+            var endIndex = GetEndOfToken(analyzer, startIndex, out var hasIntersectionWithBoldTag);
+
+            ThrowArgumentExceptionIfTokenIncorrect(analyzer, startIndex, endIndex, hasIntersectionWithBoldTag);
 
             var rawToken = mdString
                 .Substring(startIndex + MdTag.Length, endIndex - startIndex - MdTag.Length)
@@ -28,11 +32,39 @@ namespace Markdown.TokenModels
             return new ItalicToken(StringToken.Create(rawStringToken), rawStringToken.Length);
         }
 
-        private static int GetEndOfToken(string mdString, int startIndex)
+        private static void ThrowArgumentExceptionIfTokenIncorrect(
+            StringAnalyzer analyzer,
+            int startIndex,
+            int endIndex,
+            bool hasIntersectionWithBoldTag)
         {
-            var analyzer = new StringAnalyzer(mdString);
+            if (endIndex - startIndex <= 1)
+                throw new ArgumentException($"{nameof(ItalicToken)} should has length more than 0!");
+
+            if (!analyzer.HasValueUnderscoreAt(startIndex) || !analyzer.HasValueUnderscoreAt(endIndex))
+                throw new ArgumentException($"{nameof(ItalicToken)} should starts and ends with underscore!");
+
+            if (analyzer.HasValueWhiteSpaceAt(endIndex - 1))
+                throw new ArgumentException($"{nameof(ItalicToken)} shouldn't has white space before close tag!");
+
+            if (hasIntersectionWithBoldTag)
+                throw new ArgumentException(
+                    $"{nameof(ItalicToken)} should ends with underscore and shouldn't has between open and close tags unpaired double underscore"
+                );
+
+            if (IsDigitAroundItalicTag(analyzer, startIndex) || IsDigitAroundItalicTag(analyzer, endIndex))
+                throw new ArgumentException($"{nameof(ItalicToken)} shouldn't has digits around open and close tags!");
+
+            if (analyzer.HasValueSelectionPartWordInDifferentWords(startIndex, endIndex))
+                throw new ArgumentException(
+                    $"{nameof(ItalicToken)} must highlight either part of a single word, or the entire word!"
+                );
+        }
+
+        private static int GetEndOfToken(StringAnalyzer analyzer, int startIndex, out bool hasIntersectionWithBoldTag)
+        {
             var endIndex = startIndex + MdTag.Length;
-            var hasIntersectionWithBoldTag = false;
+            hasIntersectionWithBoldTag = false;
 
             while (analyzer.IsCharInsideValue(endIndex) && !analyzer.HasValueUnderscoreAt(endIndex))
             {
@@ -44,9 +76,16 @@ namespace Markdown.TokenModels
                 }
             }
 
-            TokenThrowHelper.AssertThatExtractedItalicTokenCorrect(analyzer, startIndex, endIndex,
-                hasIntersectionWithBoldTag);
             return endIndex;
+        }
+
+        private static bool IsDigitAroundItalicTag(StringAnalyzer analyzer, int position)
+        {
+            var hasDigitBefore = analyzer.IsCharInsideValue(position - 1) &&
+                                 char.IsDigit(analyzer.AnalyzedString[position - 1]);
+            var hasDigitAfter = analyzer.IsCharInsideValue(position + 1) &&
+                                char.IsDigit(analyzer.AnalyzedString[position + 1]);
+            return hasDigitBefore || hasDigitAfter;
         }
     }
 }
